FAQ: Reworded "Name Stealing" section.
The following claim was removed, because anecdotal and unsourced evidence are crap: "Sex.com was famously stolen **with a fax** and it took *5 years and $4.5 million* in legal fees to recover.  The damages awarded to the original owner have yet to be paid.  While this case occurred in the 90s, we know of national and commercial registrars that could easily tricked into transferring a domain."